requirements.txt
grpcio==1.58.0
loguru==0.7.2
passlib==1.7.4
peewee==3.16.3
protobuf==4.24.3
Requests==2.31.0
PyMySQL==1.1.0
grpcio-tools==1.58.0
python-consul2==0.1.5
